[hdr] move sampling in a dedicated file
